+++ actiontext/CHANGELOG.md
@@ -3,6 +3,8 @@
 *   Sanitize ActionText HTML ContentAttachment in Trix edit view
     [CVE-2024-32464]
 
+    *Aaron Patterson*, *Zack Deveau*
+
 
 ## Rails 7.2.0.beta1 (May 29, 2024) ##